Use common check_location_exists method
